requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests==2.28.2
six==1.16.0
sseclient-py==1.7.2
urllib3==1.26.16
Werkzeug==2.3.6
